refactor(mobile): split large mobile surfaces
